refactor: corregidas rutas a dtos en paquetes, retirado JWT_SECRET de properties
