+++ src/main/resources/application.properties
@@ -22,7 +22,9 @@ here.maps.api.url=https://discover.search.hereapi.com/v1/
 besttime.api.url=https://besttime.app/api/v1/venues/
 
 # Key para JWT todo Debe estar en variables de entorno!
-secret.key = "ZXN0YSBlcyBsYSBzZWNyZXQga2V5IHBhcmEgYWx0b3VyIGVzIG11eSBzZWNyZXRhIHkgbm8gc2UgcHVlZGUgY29tcGFydGlyIQ==";
+secret.key=${JWT_SECRET_KEY}
+
+
